Refactor linked provider and add solana address validation in token gated
