+++ src/modules/telegram/ContestEvmModal.tsx
@@ -3,7 +3,7 @@ import Button from '@/components/Button'
 import LinkText from '@/components/LinkText'
 import Input from '@/components/inputs/Input'
 import useIsMounted from '@/hooks/useIsMounted'
-import useLinkedAddress from '@/hooks/useLinkedEvmAddress'
+import useLinkedAddress from '@/hooks/useLinkedProviders'
 import { useAddExternalProviderToIdentity } from '@/services/datahub/identity/mutation'
 import { useSendEvent } from '@/stores/analytics'
 import { useMyMainAddress } from '@/stores/my-account'
