+++ src/modules/telegram/MenuPage/index.tsx
@@ -10,7 +10,7 @@ import UnlinkWalletModal from '@/components/wallets/UnlinkWalletModal'
 import UseMobileModal from '@/components/wallets/UseMobileModal'
 import EvmConnectWalletModal from '@/components/wallets/evm/EvmConnectWalletModal'
 import useIsModerationAdmin from '@/hooks/useIsModerationAdmin'
-import { useLinkedProviders } from '@/hooks/useLinkedEvmAddress'
+import { useLinkedProviders } from '@/hooks/useLinkedProviders'
 import useTgNoScroll from '@/hooks/useTgNoScroll'
 import PointsWidget from '@/modules/points/PointsWidget'
 import { useMyMainAddress } from '@/stores/my-account'
